Delete Model Scenes, Rename Things
"Push Up" and "Raise" have been unified as "Raise";
Model Scenes are functionless and have been removed.

The "Playable" Scenes compose all the models manually.
I wonder if in the outermost levels, the models must be composed then
in order to work with them. That might be tedious, but also
there probably won't be so many models to work with.

Maybe models will create their children like DynamicBlocks do with Exploders
and Fallers. Views would have to unravel in the same way too, ha.
Design is hard.